requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
et-xmlfile==1.1.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.2
gunicorn==20.1.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.24.1
openpyxl==3.0.10
packaging==23.1
pandas==1.5.2
pluggy==1.0.0
psycopg2==2.9.5
pytest==7.3.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7.1
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.44
urllib3==1.26.12
uWSGI==2.0.21
Werkzeug==2.2.2
